Fixes a bug when a node is disconnected ungracefully from a two node cluster.
It seems like a single cluster does not record itself in the leafset. So
when the leafset is repaired, it doesn't find the "next node" and
complains. But the function continues and the `send` ends up failing.
This results in the neighborhoodset never being updated/repaired.

My solution was to simply return nil when repairing an empty leafset.